Rename CoinAmount to ApiAmount; Break up Module of util/amount; Make ApiAmount the default schema/serialization
